LayoutTests/inspector-protocol/dom-request-child-nodes-depth-expected.txt


=== Get the Document ===


=== Get immediate children of the body ===

PASS: First child is a div
PASS: First child is div#depth-1
PASS: First child has one child
PASS: First child has no .children property

=== Get all children of div#depth-1 ===

PASS: div#depth-1 has nodes 9 levels deep

=== Pass an invalid depth ===

Backend error: Please provide a positive integer as a depth or -1 for entire subtree (-32000)

PASS: Expected number of setChildNodes events
